[Chore] styles -> globalStyles로 변경 (#3)
